+++ swap_test.go
@@ -1,7 +1,5 @@
 // Copyright 2026 github.com/mixcode
 
-//go:build !safe_binarystruct
-
 package binarystruct
 
 import (